+++ src/uet/oop/bomberman/graphics/Sprite.java
@@ -216,8 +216,8 @@ public class Sprite {
 		}
 	}
 	
-	public static Sprite movingSprite(Sprite normal, Sprite x1, Sprite x2, int animate, int time) {
-		int calc = animate % time;
+	public static Sprite movingSprite(Sprite normal, Sprite x1, Sprite x2, long animate, int time) {
+		long calc = animate % time;
 		int diff = time / 3;
 		
 		if(calc < diff) {
